+++ benchmarks/ops/bench_mhc_post.py
@@ -5,7 +5,7 @@ import torch
 
 from benchmarks.benchmark import BenchmarkBase, BenchmarkReport
 from tileops.ops import MHCPostOp
-from workloads.mhc_post import MHCPostTest
+from workloads.mhc import MHCPostTest
 
 
 class _MHCPostTestBaseline(MHCPostTest):
