+++ benchmarks/ops/bench_moe_permute.py
@@ -32,7 +32,7 @@ except ImportError:
 from benchmarks.benchmark import BenchmarkBase, BenchmarkReport
 from tileops.manifest import eval_roofline, load_workloads
 from tileops.ops.moe import MoePermutePaddedFwdOp
-from workloads.moe_permute import MoePermuteTest
+from workloads.moe import MoePermuteTest
 
 _OP_NAME = "MoePermutePaddedFwdOp"
