+++ tests/ops/test_gated_deltanet_recurrence.py
@@ -6,7 +6,7 @@ import torch
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import GatedDeltaNetDecodeOp
-from workloads.gated_deltanet_recurrence import (
+from workloads.gated_deltanet import (
     GatedDeltaNetDecodeTest as _GatedDeltaNetDecodeTestWorkload,
 )
