+++ tests/ops/test_gla_recurrence.py
@@ -6,7 +6,7 @@ import torch
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import GLADecodeOp
-from workloads.gla_recurrence import GLADecodeTest as _GLADecodeTestWorkload
+from workloads.gla import GLADecodeTest as _GLADecodeTestWorkload
 
 
 def gla_decode_torch(
