+++ tests/ops/test_moe_unpermute.py
@@ -18,7 +18,7 @@ import torch
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops.moe import MoeUnpermuteFwdOp
-from workloads.moe_unpermute import MoeUnpermuteTest as _MoeUnpermuteTestWorkload
+from workloads.moe import MoeUnpermuteTest as _MoeUnpermuteTestWorkload
 
 
 def _ref_moe_unpermute(
