--- workloads/deltanet_bwd.py
@@ -1,3 +1,0 @@
-# This module previously contained oracle/reference functions (_differentiable_fwd,
-# deltanet_autograd_bwd_torch). Those have been moved to their respective test and
-# benchmark files as part of the trust-model separation (no shared oracle surface).
