--- workloads/deltanet_recurrence.py
@@ -1,31 +1,0 @@
-from typing import Tuple
-
-import torch
-
-from workloads.base import WorkloadBase
-
-
-class DeltaNetDecodeTest(WorkloadBase):
-
-    def __init__(
-        self,
-        batch: int,
-        heads: int,
-        dim_k: int,
-        dim_v: int,
-        dtype: torch.dtype,
-    ) -> None:
-        self.batch = batch
-        self.heads = heads
-        self.dim_k = dim_k
-        self.dim_v = dim_v
-        self.dtype = dtype
-
-    def gen_inputs(self) -> Tuple[torch.Tensor, ...]:
-        B, H, DK, DV = self.batch, self.heads, self.dim_k, self.dim_v
-        q = torch.randn(B, H, DK, device="cuda", dtype=self.dtype) * 0.1
-        k = torch.randn(B, H, DK, device="cuda", dtype=self.dtype) * 0.1
-        v = torch.randn(B, H, DV, device="cuda", dtype=self.dtype) * 0.1
-        beta = torch.rand(B, H, device="cuda", dtype=self.dtype) * 0.5
-        state = torch.randn(B, H, DK, DV, device="cuda", dtype=self.dtype) * 0.1
-        return q, k, v, beta, state
