+++ workloads/moe.py
@@ -1,0 +1,79 @@
+from typing import Tuple
+
+import torch
+
+from workloads.base import WorkloadBase
+
+
+class FusedTopKTest(WorkloadBase):
+    def __init__(self, num_tokens, num_experts, top_k, scoring_func, renormalize, dtype):
+        self.num_tokens = num_tokens
+        self.num_experts = num_experts
+        self.top_k = top_k
+        self.scoring_func = scoring_func
+        self.renormalize = renormalize
+        self.dtype = dtype
+
+    def gen_inputs(self):
+        torch.manual_seed(42)
+        return torch.randn(self.num_tokens, self.num_experts, dtype=self.dtype, device="cuda")
+
+
+class MoePermuteTest(WorkloadBase):
+
+    def __init__(self, total_tokens, top_k, num_experts, hidden_size, dtype):
+        self.total_tokens = total_tokens
+        self.top_k = top_k
+        self.num_experts = num_experts
+        self.hidden_size = hidden_size
+        self.dtype = dtype
+
+    def gen_inputs(self):
+        hidden_states = torch.randn(
+            self.total_tokens, self.hidden_size, dtype=self.dtype, device="cuda"
+        )
+        topk_ids = torch.randint(
+            0, self.num_experts,
+            (self.total_tokens, self.top_k),
+            dtype=torch.int32, device="cuda",
+        )
+        return hidden_states, topk_ids
+
+
+class MoePermuteAlignTest(WorkloadBase):
+
+    def __init__(self, total_tokens: int, top_k: int, num_experts: int, block_size: int):
+        self.total_tokens = total_tokens
+        self.top_k = top_k
+        self.num_experts = num_experts
+        self.block_size = block_size
+
+    def gen_inputs(self) -> Tuple[torch.Tensor]:
+        topk_ids = torch.randint(
+            0, self.num_experts,
+            (self.total_tokens, self.top_k),
+            dtype=torch.int32, device="cuda",
+        )
+        return (topk_ids,)
+
+
+class MoeUnpermuteTest(WorkloadBase):
+
+    def __init__(self, total_tokens, top_k, hidden_size, dtype):
+        self.total_tokens = total_tokens
+        self.top_k = top_k
+        self.hidden_size = hidden_size
+        self.dtype = dtype
+        # Use padded_batch_sum = T*K (no actual padding) for standalone tests.
+        self.padded_batch_sum = total_tokens * top_k
+
+    def gen_inputs(self):
+        numel = self.total_tokens * self.top_k
+        mm2_pad = torch.randn(numel, self.hidden_size, dtype=self.dtype, device="cuda")
+        # fwd_idx: each flat_idx maps to a padded_slot in [0, padded_batch_sum)
+        # simulate a valid mapping: random shuffle of [0, numel)
+        fwd_idx = torch.randperm(numel, dtype=torch.int32, device="cuda")
+        topk_weights = torch.rand(
+            self.total_tokens, self.top_k, dtype=torch.float32, device="cuda"
+        )
+        return mm2_pad, fwd_idx, topk_weights
