--- workloads/moe_permute.py
@@ -1,24 +1,0 @@
-import torch
-
-from workloads.base import WorkloadBase
-
-
-class MoePermuteTest(WorkloadBase):
-
-    def __init__(self, total_tokens, top_k, num_experts, hidden_size, dtype):
-        self.total_tokens = total_tokens
-        self.top_k = top_k
-        self.num_experts = num_experts
-        self.hidden_size = hidden_size
-        self.dtype = dtype
-
-    def gen_inputs(self):
-        hidden_states = torch.randn(
-            self.total_tokens, self.hidden_size, dtype=self.dtype, device="cuda"
-        )
-        topk_ids = torch.randint(
-            0, self.num_experts,
-            (self.total_tokens, self.top_k),
-            dtype=torch.int32, device="cuda",
-        )
-        return hidden_states, topk_ids
