--- workloads/moe_permute_align.py
@@ -1,22 +1,0 @@
-from typing import Tuple
-
-import torch
-
-from workloads.base import WorkloadBase
-
-
-class MoePermuteAlignTest(WorkloadBase):
-
-    def __init__(self, total_tokens: int, top_k: int, num_experts: int, block_size: int):
-        self.total_tokens = total_tokens
-        self.top_k = top_k
-        self.num_experts = num_experts
-        self.block_size = block_size
-
-    def gen_inputs(self) -> Tuple[torch.Tensor]:
-        topk_ids = torch.randint(
-            0, self.num_experts,
-            (self.total_tokens, self.top_k),
-            dtype=torch.int32, device="cuda",
-        )
-        return (topk_ids,)
